Fix: Resolve all compiler warnings

+++ adc.c
@@ -2,32 +2,32 @@
 #include "adc.h"
 
 void ADC_Init(void) {
-    // ADCON1: Configura pinos analógicos e VRef
-    // PCFG<3:0> = 1101: AN0 e AN1 são analógicos, os demais são digitais.
-    // VCFG<1:0> = 00: Tensão de referência VSS e VDD.
+    // ADCON1: Configura pinos analï¿½gicos e VRef
+    // PCFG<3:0> = 1101: AN0 e AN1 sï¿½o analï¿½gicos, os demais sï¿½o digitais.
+    // VCFG<1:0> = 00: Tensï¿½o de referï¿½ncia VSS e VDD.
     ADCON1 = 0x0D; 
 
-    // ADCON2: Configura justificação, tempo de aquisição e clock do ADC
-    // ADFM = 1: Resultado justificado à direita (bits 9:0 em ADRESH:ADRESL)
-    // ACQT<2:0> = 101: 12 TAD (Tempo de Aquisição)
+    // ADCON2: Configura justificaï¿½ï¿½o, tempo de aquisiï¿½ï¿½o e clock do ADC
+    // ADFM = 1: Resultado justificado ï¿½ direita (bits 9:0 em ADRESH:ADRESL)
+    // ACQT<2:0> = 101: 12 TAD (Tempo de Aquisiï¿½ï¿½o)
     // ADCS<2:0> = 010: Clock FOSC/32
     ADCON2 = 0xB5; // 0b10110101
     
-    // ADCON0: Habilita o módulo ADC
+    // ADCON0: Habilita o mï¿½dulo ADC
     ADCON0bits.ADON = 1;
 }
 
 unsigned int ADC_Read(unsigned char channel) {
     // 1. Seleciona o canal
-    // Limpa os bits do canal (CHS<3:0>) e mantém ADON ligado
+    // Limpa os bits do canal (CHS<3:0>) e mantï¿½m ADON ligado
     ADCON0 = (ADCON0 & 0xC3) | ((channel << 2) & 0x3C);
 
-    // 2. Inicia a conversão
+    // 2. Inicia a conversï¿½o
     ADCON0bits.GO_DONE = 1;
 
-    // 3. Espera a conversão terminar
+    // 3. Espera a conversï¿½o terminar
     while (ADCON0bits.GO_DONE);
 
     // 4. Retorna o resultado (10 bits)
-    return ((ADRESH << 8) | ADRESL);
+    return ((unsigned int)ADRESH << 8) | (unsigned int)ADRESL;
 }
